admin_service/requirements.txt
django==4.1.7
flake8==6.1.0
python-dotenv==0.21.0
django-split-settings==1.2.0
django-debug-toolbar==4.3.0
psycopg2_binary==2.9.9
gunicorn==22.0.0
StrEnum==0.4.15
requests==2.31.0